+++ pkg/services/auth/idimpl/service.go
@@ -11,6 +11,8 @@ import (
 	"github.com/grafana/grafana/pkg/infra/remotecache"
 	"github.com/grafana/grafana/pkg/services/auth"
 	"github.com/grafana/grafana/pkg/services/auth/identity"
+	"github.com/grafana/grafana/pkg/services/authn"
+	"github.com/grafana/grafana/pkg/services/featuremgmt"
 	"github.com/grafana/grafana/pkg/setting"
 )
 
@@ -22,8 +24,17 @@ const (
 
 var _ auth.IDService = (*Service)(nil)
 
-func ProvideService(cfg *setting.Cfg, signer auth.IDSigner, cache remotecache.CacheStorage) *Service {
-	return &Service{cfg, log.New("id-service"), signer, cache}
+func ProvideService(
+	cfg *setting.Cfg, signer auth.IDSigner, cache remotecache.CacheStorage,
+	features featuremgmt.FeatureToggles, authnService authn.Service,
+) *Service {
+	s := &Service{cfg, log.New("id-service"), signer, cache}
+
+	if features.IsEnabled(featuremgmt.FlagIdForwarding) {
+		authnService.RegisterPostAuthHook(s.hook, 140)
+	}
+
+	return s
 }
 
 type Service struct {
@@ -46,7 +57,6 @@ func (s *Service) SignIdentity(ctx context.Context, id identity.Requester) (stri
 	s.logger.Debug("Sign new id token", "namespace", namespace, "id", identifier)
 
 	now := time.Now()
-
 	token, err := s.signer.SignIDToken(ctx, &auth.IDClaims{
 		Claims: jwt.Claims{
 			ID:       identifier,
@@ -67,6 +77,21 @@ func (s *Service) SignIdentity(ctx context.Context, id identity.Requester) (stri
 	}
 
 	return token, nil
+}
+
+func (s *Service) hook(ctx context.Context, identity *authn.Identity, _ *authn.Request) error {
+	// FIXME(kalleep): implement identity.Requester for authn.Identity
+	// FIXME(kalleep): we should probably lazy load this
+	token, err := s.SignIdentity(ctx, identity.SignedInUser())
+	if err != nil {
+		namespace, id := identity.NamespacedID()
+		s.logger.Error("Failed to sign id token", "err", err, "namespace", namespace, "id", id)
+		// for now don't return error so we don't break authentication from this hook
+		return nil
+	}
+
+	identity.IDToken = token
+	return nil
 }
 
 func prefixCacheKey(key string) string {
